# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/devices": {
            "get": {
                "description": "Get a list of Devices.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Get all Devices.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Devices"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new Devices.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Create New Devices.",
                "parameters": [
                    {
                        "description": "the body to create a new Devices",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.devicesInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Devices"
                        }
                    }
                }
            }
        },
        "/devices/{id}": {
            "get": {
                "description": "Get an Devices by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Get Devices.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Devices id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Devices"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a Devices by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Delete one Devices.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Devices id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update Devices by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Update Devices.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Devices id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update devices",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.devicesInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Devices"
                        }
                    }
                }
            }
        },
        "/devices/{id}/comments": {
            "get": {
                "description": "Get all Opinion by DevicesId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Get Comments By id Devices.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Devices id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Comments"
                            }
                        }
                    }
                }
            }
        },
...
        "/opinions": {
            "get": {
                "description": "Get a list of Opinions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opinions"
                ],
                "summary": "Get all Opinions.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Opinions"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Opinions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opinions"
                ],
                "summary": "Create New Opinions.",
                "parameters": [
                    {
                        "description": "the body to create a new Opinions",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.opinionsInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Opinions"
                        }
                    }
                }
            }
        },
        "/opinions/{id}": {
            "get": {
                "description": "Get an Opinions by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opinions"
                ],
                "summary": "Get Opinions.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opinions id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Opinions"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a Opinions by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opinions"
                ],
                "summary": "Delete one Opinions.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opinions id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update Opinions by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Opinions"
                ],
                "summary": "Update Opinions.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opinions id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update opinions",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.opinionsInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Opinions"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.devicesInput": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "specification_id": {
                    "type": "integer"
                }
            }
        },
        "controller.opinionsInput": {
            "type": "object",
            "properties": {
                "opinion_section": {
                    "type": "string"
                },
                "reviews_id": {
                    "type": "integer"
                },
                "your_name": {
                    "type": "string"
                }
            }
        },
...
        "models.Comments": {
            "type": "object",
            "properties": {
                "comments_section": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "devicesID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "reviewsID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "your_name": {
                    "type": "string"
                }
            }
        },
        "models.Devices": {
            "type": "object",
            "properties": {
                "brandID": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "specificationID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Opinions": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "opinion_section": {
                    "type": "string"
                },
                "reviewsID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "your_name": {
                    "type": "string"
                }
            }
        },
...
    }
}